requirements.txt
contourpy==1.3.3
cycler==0.12.1
filelock==3.19.1
fonttools==4.60.1
fsspec==2025.9.0
git-filter-repo==2.47.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.7
mpmath==1.3.0
networkx==3.5
numpy==2.3.3
packaging==25.0
pandas==2.3.3
pillow==11.3.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytorch-triton-rocm==3.5.0+git7416ffcb
pytz==2025.2
setuptools==70.2.0
six==1.17.0
sympy==1.14.0
tabulate==0.9.0
torch==2.10.0.dev20251028+rocm7.0
torchvision==0.25.0.dev20251029+rocm7.0
typing_extensions==4.15.0
tzdata==2025.2
